Automatically fallback to default_settings.toml
Previously if user_settings.toml was not present then a prompt asking
for explicit confirmation for using default_settings.toml would appear,
requiring manual input.

This changes the script to automatically use default_settings.toml if
user_settings.toml does not exist.

Additionally, it causes the script to exit if default_settings.toml
is to be fallen back to but it also does not exit - it should always
exist as it's checked into the repo and the script would previously
have simply crashed at another point further along.